Fraud protection
We recently learned that some criminals are directly selling the source code of open source projects.
Since this behavior does not violate the GPL agreement,
We only provide copyright information to inform users that this project is free software.
Please note:
Free software does not mean that this is free software, and this project can be used for commercial purposes.
However, the copyright belongs to DESMG, and DESMG reserves all rights.
We chose the GPL because we want this project to be free forever.